saving to localStorage()
still need work on individual items to be saved properly

got it where it would save each item on reload by using the forEach method, passing it through a new function and calling it through an event listener "DOMContentLoaded". Utilizing the OR operator (||) so no error resides on reload as JS will try to parse an empty array.